refactor(debug): split overlay editor module responsibilities
