+++ Plugins/PortfolioDebugOverlayEditor/Source/PortfolioDebugOverlayEditor/Private/FPortfolioDebugOverlayEditorFocusCommandBridge.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "CoreMinimal.h"
+
+namespace PortfolioDebugOverlayEditorFocusCommandBridge
+{
+	FText ExecuteSelectNearestFocusCommand();
+	FText ExecuteSelectRecentCombatFocusCommand();
+	FText ExecuteClearFocusCommand();
+	FText ExecuteSelectOutlinerFocusCommand();
+}
